Stop using Flow on tasks not using Progress
Flow forces to use FlowContext (bad thing) and apply progress (good thing). So if the progress is not used, it just force you to use FlowContext. So just do it with pure RX using Completable.andThen. This also should make it easy to try to migrate from FlowContext to type-safe RX chain in those places (bc progress add a it of complexity, so better to try out without progress involved first).